+++ src/main/java/com/tagnumelite/projecteintegration/PEIntegration.java
@@ -1,8 +1,12 @@
 package com.tagnumelite.projecteintegration;
 
+import com.tagnumelite.projecteintegration.api.recipe.PEIRecipeMapper;
 import moze_intel.projecte.PECore;
 import moze_intel.projecte.config.ProjectEConfig;
+import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,6 +16,8 @@ public class PEIntegration {
     public static final Logger LOGGER = LogManager.getLogger(MODID);
 
     public PEIntegration() {
+        IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
+        modEventBus.addListener(this::commonSetup);
     }
 
     public static void debugLog(String msg, Object... args) {
@@ -22,4 +28,7 @@ public class PEIntegration {
         }
     }
 
+    private void commonSetup(FMLCommonSetupEvent event) {
+        PEIRecipeMapper.loadMappers();
+    }
 }